# services/backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: SkoLab API
  version: "1.0"
  description: |
    The public API of skolab-api: the Go gateway. (The Python service runs in
    the same container and is not reachable from outside; compiles reach it
    through POST /api/v1/colab/compile.)

    Every error is `{"error": "<message for people>", "code": "<stable code>"}`;
    clients branch on `code`. Every response carries HSTS, nosniff and frame
    and referrer policies, and responses under /api are never cached, except
    the template catalog, which a browser may keep but must revalidate (ETag).

    The collaboration WebSocket (`GET /ws/colab/{workspace_id}?ticket=...`)
    is not described here: open it with a ticket from
    `POST /api/v1/ws/colab/{workspace_id}/tickets`.

    This file is the contract the QA tools test against (services/qa).
servers:
  - url: https://skolab-api.onrender.com
security:
  - firebase: []
...
paths:
...
  /api/v1/workspaces:
    post:
      tags: [workspaces]
      operationId: createWorkspace
      summary: Create a workspace (the caller becomes owner)
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Retrying with the same key and body returns the same workspace (201, Idempotent-Replayed true). Empty means no key. At most 100 printable ASCII characters; trailing spaces or tabs are removed in transit, as for any HTTP header.
          schema: { type: string, pattern: "^([!-~]{1,100}[ \t]*)?$" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Title" }
      responses:
        "201":
          description: Created (or replayed).
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "422":
          description: Idempotency-Key already used for a different request.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
    get:
      tags: [workspaces]
      operationId: listWorkspaces
      summary: Workspaces the caller owns or is a member of
      description: Query parameters other than page_size and page_token are refused (400 unknown_parameter), and so is any parameter given more than once (400 duplicate_parameter).
      parameters:
        - { name: page_size, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 100 } }
        - { name: page_token, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: One page.
          content:
            application/json:
              schema:
                type: object
                required: [workspaces, next_page_token]
                properties:
                  workspaces: { type: array, items: { $ref: "#/components/schemas/Workspace" } }
                  next_page_token: { type: string, description: Empty on the last page. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/workspaces/{id}/files:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [editor]
      operationId: listFiles
      summary: The project's files and folders, main.tex and the last compiled PDF (any active member)
      responses:
        "200":
          description: The project at a glance, files sorted by path.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileListing" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
    post:
      tags: [editor]
      operationId: createFile
      summary: Create a text file or a folder (owner or editor)
      description: |
        Folders above the path are created as needed. Paths are relative,
        "/"-separated, unique ignoring case, at most 5 folders deep; main.tex
        and output.pdf at the top are reserved. A project holds at most 10 MiB
        in 200 entries (413 project_full).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileCreate" }
      responses:
        "201":
          description: Created. A text file comes with its content.
          headers:
            ETag: { schema: { type: string, pattern: '^"v[0-9]+"$' } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectFile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "415": { $ref: "#/components/responses/Unsupported" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/workspaces/{id}/files/upload:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [editor]
      operationId: uploadFiles
      summary: Upload files into the project (owner or editor)
      description: |
        One or more `file` parts (at most 5 MiB each, 12 MiB in all). Text
        files (.tex .bib .bst .cls .sty .txt .md .csv and other TeX files) must
        be UTF-8; images and PDFs (.png .jpg .jpeg .pdf .eps) must look like
        what their extension says (415 unsupported_file_type). `folder` puts
        them in a folder; `replace=true` overwrites files with the same path
        (409 file_exists otherwise).
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: array
                  items: { type: string, format: binary }
                folder: { type: string, maxLength: 200 }
                replace: { type: string, enum: ["true", "false"] }
      responses:
        "201":
          description: Stored, in upload order.
          content:
            application/json:
              schema:
                type: object
                required: [files]
                properties:
                  files: { type: array, items: { $ref: "#/components/schemas/ProjectFile" } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "415": { $ref: "#/components/responses/Unsupported" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/workspaces/{id}/files/{file_id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/FileId"
    get:
      tags: [editor]
      operationId: getFile
      summary: One file or folder; a text file comes with its content (any active member)
      responses:
        "200":
          description: The file.
          headers:
            ETag: { schema: { type: string, pattern: '^"v[0-9]+"$' } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectFile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
    put:
      tags: [editor]
      operationId: saveFile
      summary: Save a text file (owner or editor)
      description: Like saving the document, base_version must be the current version, else 409 version_conflict.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FileSave" }
      responses:
        "200":
          description: Saved; version is base_version + 1.
          headers:
            ETag: { schema: { type: string, pattern: '^"v[0-9]+"$' } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectFile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: The file changed since base_version.
          headers:
            ETag: { schema: { type: string, pattern: '^"v[0-9]+"$' } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionConflict" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
    patch:
      tags: [editor]
      operationId: moveFile
      summary: Rename or move a file or folder; a folder takes its contents along, a file keeps its type (owner or editor)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              additionalProperties: false
              properties:
                path: { $ref: "#/components/schemas/FilePath" }
      responses:
        "200":
          description: Moved.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProjectFile" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "415": { $ref: "#/components/responses/Unsupported" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
    delete:
      tags: [editor]
      operationId: deleteFile
      summary: Delete a file, or a folder with everything in it (owner or editor)
      responses:
        "204": { description: Deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/workspaces/{id}/archive:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [editor]
      operationId: downloadProject
      summary: The whole project as a .zip (any active member)
      description: main.tex, every file and folder, and output.pdf when the project has been compiled.
      responses:
        "200":
          description: The archive, named after the workspace.
          headers:
            Content-Disposition: { schema: { type: string } }
          content:
            application/zip: { schema: { type: string, format: binary } }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
components:
  securitySchemes:
    firebase:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    sentrySignature:
      type: apiKey
      in: header
      name: Sentry-Hook-Signature
  parameters:
    WorkspaceId:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Title:
      type: object
      required: [title]
      properties:
        title: { $ref: "#/components/schemas/Text255" }
...
    Workspace:
      type: object
      required: [id, title, owner_id, role, created_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        owner_id: { type: string }
        role: { $ref: "#/components/schemas/Role" }
        created_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: The request is malformed or outside the contract.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...